break the parse file into seperate functions and store it in src
